Fix smartrap.jscad file
This fixes the JSCAD file and translates all objects to have a ready-to-print STL file.

Unnecessary variables and methods (that were used nowhere in the code) are removed, the result is now always a static array containing all parts of the file. As only the xend_ prefixed parts were included, those are now also on the final plate.